Moving the container modification logic in the ItemContainer
When creating an Item, if `container_id`, `container_class`,
`container_attributes` or `selected_class` are provided in options,
the Item would modify its own ItemContainer.

This commit moves this logic in ItemContainer itself. When adding
an item to a container using `item` or `items=` the logic will
be called the same way.

Creating a standalone Item shouldn't have side effects on its
container, only explicitly adding it to a container should have
this impact.