Remove query string to determine active link
When viewing that require a navigation (with a query string ?page), the active link stops working. I suggest that we add the possibility to allow the user to define: `include_query: true` in the item definition to revert to the old behaviour. Else, the query is always omitted from the url.